Beefed up combat in TSF
I upgraded the combat in TSF. The reinforcement wave of ICE are now
using the lvl 6 ESP Attacker, Kamikaze, and Shield models and character
sheets, and they can't be killed, but they can be injured. I'll add a
trigger that notifies the player that they cannot kill them, and should
just run.

Current plans are to update the other IC in the level, and move
Armitage's "go on without me" quote to go off just before the player
leaves TSF