Version 1.1
A detailed rundown of the changes in this patch will be written up, and
posted on the Development Blog. So, if you're interested in learning my
thought process behind these patch changes, stay tuned to the Project
ShadowMancer blog!

- Improved The Chase Scenes. Molly should now actually pose a threat.

-Case has a basic attack during these chases after he acquires the
cobra, but it seems a bit off...

-The combat in The Screaming Fist has been improved! The enemies in the
beginning should pose more of a threat now.